unittest and parse.py file
